Add inline PlantUML support for markdown files
It's now possible to write PlantUML inline in a markdown file by using
the standard markdown syntax for PlantUML. The render-plantuml.sh script
is used to generate images based on these inline PlantUML blocks.

After rendering, it is possible to refer to the images directly in
markdown. The images will be named the same as the string directly after
the start of the UML with the `@startuml` part.

This is currently a manual step, but can be expanded to run in GitHub
Actions or something similar.